Add configuration key for enabling (disabling) agent proxy telemetry
This is primarily useful in testing scenarios, but it also allows us to skip trying to send to the agent when we're in scenarios we know we don't have an agent, or where we _probably_ don't have an agent. The latter is the case for CIApp when in agentless mode, and may well become more prevalent in that future.